Move isFilterObject type guard
